Stored the existing UI in a folder named "Draft UI" and made the Poll UI the default UI
