+++ Lost-and-Found-App/src/Draft-UI/App.tsx
@@ -1,0 +1,7 @@
+import AppDefault from "./AppDefault";
+import AppUITest from "./AppUITest";
+
+export default function App() {
+  const useUiTest = import.meta.env.VITE_UI_TEST === "true";
+  return useUiTest ? <AppUITest /> : <AppDefault />;
+}
