+++ Lost-and-Found-App/src/Draft-UI/index.css
@@ -4,8 +4,8 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  background-color: #ffffff;
-   color: #111;
+  color: rgba(255, 255, 255, 0.87);
+  background-color: #242424;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
@@ -24,13 +24,10 @@ a:hover {
 
 body {
   margin: 0;
+  display: flex;
+  place-items: center;
   min-width: 320px;
   min-height: 100vh;
-
-  background: #e6e6e6;
-
-  display: flex;
-  flex-direction: column;
 }
 
 h1 {
