--- Lost-and-Found-App/src/UI-Poll/App.tsx
@@ -1,5 +1,0 @@
-import HomePage from "./pages/HomePage";
-
-export default function App() {
-  return <HomePage />;
-}
